v2/api/pyproject.toml
[project]
name = "lalagolf-api"
version = "0.1.0"
description = "FastAPI backend for LalaGolf v2"
requires-python = ">=3.12"
dependencies = [
  "alembic==1.15.2",
  "fastapi==0.115.12",
  "uvicorn[standard]==0.34.0",
  "pydantic-settings==2.8.1",
  "python-multipart==0.0.20",
  "sqlalchemy==2.0.40",
  "psycopg[binary]==3.2.6",
  "redis==5.2.1"
]

[project.optional-dependencies]
dev = [
  "httpx==0.28.1",
  "pytest==8.3.5",
  "ruff==0.11.5"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
